feature: added a button for data-download to map-performance, for convenience
