+++ projects/demo-maps/src/app/route-components/route-example-olperformance/fetchbutton/fetchbutton.component.html
@@ -1,0 +1,21 @@
+<div *ngIf="state$ | async as state" style="margin: 10px;">
+
+    <button *ngIf="state == 'ready'" class="btn btn-primary" (click)="fetchClicked()">
+        Fetch data
+    </button>
+
+    <button *ngIf="state == 'fetching'" class="btn btn-primary">
+        Fetching ... 
+        <span class="spinner spinner-inline spinner-sm">
+            Fetching ...
+        </span>
+    </button>
+
+    <button *ngIf="state == 'done'" class="btn btn-primary" disabled>
+        Data loaded.
+    </button>
+
+
+</div>
+
+
