+++ projects/demo-maps/src/app/route-components/route-example-olperformance/route-map7.component.html
@@ -6,6 +6,10 @@
 
     <ukis-performance></ukis-performance>
 
+    <app-fetchbutton></app-fetchbutton>
+
+    <p>Load data and pan the map for a while to see a drop in performance.</p>
+
     <clr-vertical-nav-group [clrVerticalNavGroupExpanded]="true" class="layers">
         <clr-icon shape="layers" clrVerticalNavIcon title="layers"></clr-icon>
         Layers
